Milestone 4 in progress

+++ ITC/notes-app/src/Components/Note.js
@@ -1,3 +1,4 @@
+import { useState } from "react";
 import Col from "react-bootstrap/Col";
 import moment from "moment";
 import CloseButton from "react-bootstrap/CloseButton";
@@ -8,13 +9,20 @@ const Note = (props) => {
 
   const createdDate = () => moment().format("LLL");
 
+  const [popUpModal, setPopUpModal] = useState(false);
+
+  function showModal() {
+    setPopUpModal(true);
+  }
+
   return (
-    <Col md={3} className="note-component">
+    <Col onClick={showModal} md={3} className="note-component">
       <CloseButton onClick={() => deleteNote(note.id)} />
       <div className="created-on-date">{createdDate()}</div>
       <div className="note-title">{note.title}</div>
       <div className="note-text"> {note.text}</div>
-      <NoteModal note={note} />
+
+      <NoteModal note={note} isModalOpen={popUpModal} />
     </Col>
   );
 };
